Fix EM27Sensors tests for refactor
